+++ CHANGELOG.md
@@ -6,6 +6,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## 1.1.411
+
+### Added
+ - Enhancements in surge alert notifications
+ - New content-visibility level (IFRC_NS: FR, Proj, event)
+ - Longer date interval for appeal doc scraper
+ - Bump up elasticsearch to 6.3.0, urllib3 to 1.26.7
+ - Fixing .dockerignore, bump up pip to 21.1
+ - Search parameter for most of /docs endpoints
+ - Handle Stand-Downs on Surge page
+ - GeneralDocument for general document upload
+ - Local development env change: psql 9.6 to 11
+ - Molnix_id as read only field
+ - Use molnix_id instead of pk while marking deployments inactive
+
 ## 1.1.410
 
 ### Added
@@ -815,7 +830,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
  - Fix backend/frontend pointing issue
- - Nicing *success.html and it's email_context
+ - Nicing ...success.html and it's email_context
  - Changing base_url to frontend_url in recover_password
  - Changing hardcoded dsgo... URL-s in notifications (except logo)
 
@@ -1832,7 +1847,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## 0.1.20
 
-[Unreleased]: https://github.com/IFRCGo/go-api/compare/1.1.410...HEAD
+[Unreleased]: https://github.com/IFRCGo/go-api/compare/1.1.411...HEAD
+[1.1.411]: https://github.com/IFRCGo/go-api/compare/1.1.410...1.1.411
 [1.1.410]: https://github.com/IFRCGo/go-api/compare/1.1.409...1.1.410
 [1.1.409]: https://github.com/IFRCGo/go-api/compare/1.1.408...1.1.409
 [1.1.408]: https://github.com/IFRCGo/go-api/compare/1.1.407...1.1.408